linebot-project/requirements.txt
echo "line-bot-sdk
fastapi
uvicorn
openai
langchain
chromadb
python-multipart
python-dotenv
gunicorn" > requirements.txt